chore: Working on the build
- Gulp passing until compile now
- Working on how to get everything together in one place
