+++ typescript/vrotsc/package.json
@@ -25,7 +25,7 @@
     "@types/js-yaml": "3.12.1",
     "@types/minimatch": "3.0.5",
     "@types/minimist": "1.2.0",
-    "@types/node": "^16.0.0",
+    "@types/node": "^14.0.0",
     "@types/uuid": "3.4.6",
     "@types/xmldoc": "1.1.4",
     "ansi-colors": "4.1.1",
